EDMT-336 refactor : write-record e2e test 코드 리팩토링
